pyproject.toml
[project]
name = "mkit-otoplus-providerapi"
version = "0.1.0"
description = "wrapper untuk provider api dari otoplus developer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "loguru-config>=0.1.0",
    "pydantic-settings>=2.11.0",
]
